requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
configparser==5.0.2
dropbox==11.4.1
Flask==1.1.2
greenlet==1.0.0
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mss==6.1.0
Pillow==8.1.2
ply==3.11
requests==2.25.1
six==1.14.0
SQLAlchemy==1.4.9
stone==3.2.1
urllib3==1.26.4
Werkzeug==1.0.1
XlsxWriter==1.3.9
